Move editor inspector classes out of ruleset namespace
